fix: use utc clock within container
This fixes the issue where the time wihtin the VM was
the same as the host. This lead to issue when configuring
the host timezone. Example I use Europe/Brussels on my host,
if I set the guest to Europe/Brussels time would be 1 hour ahead.
And it's not easy to fix this because every reboot the problem
would come up again. Setting the clock to UTC, makes more sense
and fixes this issue.